Use `polling` crate, per @papertigers
Mr. Tigers suggested that https://crates.io/crates/polling might be
worth a look. Using this to watch for new connections, then allowing
rayon to handle individual requests in a blocking fashion -- looking at
creating a shitload of threads each with a tiny stack size, since the
real work will be done on the other side of the door.