Working towards passing a door over a UDS
I want to be able to create a door in a child process and pass a
descriptor to it back to the parent process -- assuming the parent
process was invoked as the result of a door_call, it should be able to
send the descriptor to the child's door back to the client. The client
will then be able to invoke the newly created child door.